Rename NLP class to ModuleSelector
This better reflects the functionality of the class, as its
purpose is to select the correct Automate module and handle
piping, multiple modules etc.
