+++ examples/remove_schedule_demo.py
@@ -5,13 +5,13 @@ sys.path.append("..")
 
 from lib import Error  # noqa: E402
 from lib.automate import Automate  # noqa: E402
-from lib.nlp import nlp  # noqa: E402
+from lib.selector.selector import ModuleSelector  # noqa: E402
 from datetime import datetime, timedelta  # noqa: E402
 from lib.settings import load_settings  # noqa: E402
 
 load_settings()
 automate = Automate()
-n = nlp.NLP("en_rpa_simple", "en_core_web_md")
+n = ModuleSelector("en_rpa_simple", "en_core_web_md")
 
 # Enable the CalDav server before running the following
 try:
