+++ lib/selector/model_installer.py
